sink(ticdc): split config
